refactor(support): move ViewBinding to its own file for PSR-4 autoloading
ViewBindingRegistry.php declared both ViewBinding and ViewBindingRegistry.
Under this package's PSR-4 mapping, requiring ShieldCI\Support\ViewBinding
directly fatals unless ViewBindingRegistry happened to be loaded first -
tests only passed by accident of construction order.
